Выполнил разбиение кода на модули.
